refactor: migration from CRA, NPM to Vite, Yarn
